requirements.txt
Django~=3.2
djangorestframework
django-filter
django-model-utils
django-extensions
pytest
black
isort
pre-commit
djangorestframework-jwt
django-ordered-model
drf-yasg
python-jose
cryptography
pytest-django
factory_boy
django-environ
psycopg2-binary
celery
Redis
django-invitations
django-annoying
django-taggit
pytest-mock
django-celery-email
